Add motor encoder for elevator lift and early implementation of preset elevator heights
The encoder is configured to provide the elevators height travelled
in inches with the function getDistance(). Note that "0" will always
initially be the position of the elevator on robot startup. So in a
competition setting, it corresponds to the elevator's position when
fully lowered since that's how we intend the elevator to be positioned
at startup.